requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-genai>=1.0.0
networkx>=3.3
spacy>=3.7.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0
